requirements.txt
redis==5.0.1
codecarbon==2.3.4
torch==2.10.0
torchvision==0.25.0
streamlit==1.41.0
requests==2.31.0
pytest==8.1.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
Pillow==10.4.0